Updated migration to Oracle script
The migration script will execute migration in parallel for each symbol
available in the source database. It will take advantage of python
multiprocessing module and speed up the procress by roughly 3 times.
Note that you can enforce a cap on the max number of parallel processes
to be executed. The actual number of parallel processes will be the
minimum between the cap value and the number of parallel threads
available on the runtime platform (dynamically evaluated)